RUBY-889 Warn if dev mode installed in production
This commit will output a strong warning if it finds that we have
developer_mode enabled in production.

The check is at the point of auto-injecting the middleware for Rails,
which is the primary case where we can safely get the env.

I also made the choice to whitelist this to only warning on the
"production" environment rather than only allowing it for "development"
and "test". This seemed wiser given that we don't know other people's
env setup, and I've seen plenty of places with special dev
configurations that I wouldn't want us to spew warnings about.