RUBY-1297: Prevent shutdown from raising, fix flaky test.
If shutdown was called before the agent started we would previously raise an
error, this change instead logs that information.

It also resolves some flaky test issues by ensuring that the agent is started
in each of the tests, rather than depending on some other test having already
started the agent.

This change also resolves the issue described in RUBY-1346.