RUBY-1335 Warn on OkJson usage
If we end up in a state where we're using the OkJson JSON parser, warn
that it's happening since there's a significant performance drag to that
(why we didn't do full JSON marshalling to begin with on 1.8.7).

Currently this only happens if you explicitly configure your 1.8.7 app
to set marshaller: json, but with upcoming changes there might be corner
cases and load problems where this is a possible edge case, hence the
warning.