RUBY-1335 Test fixes for JSON on 1.8.7
This commit deals with a variety of testing weirdness that we had around
the different Ruby vs JSON formats data could be sent in. Largely there
were places that assumed that Ruby version implied not JSON format, so
those tests have been updated.

Things are still broken, though, regarding encoding for 1.8.7, but I
will handle that as a separate commit.