RUBY-1335 Don't normalize by default on 1.8.7
1.8.7 doesn't support fast lookups for encoding validity like is
provided in 1.9.x and later, so we have to either not normalize or
always convert all strings that are going through the JSON
serialization.

We're choosing to default to NOT normalize on 1.8.7 (the opposite of
what we do on 1.9.x), but the config setting can be used explicitly to
force normalization if there's a reason to.