RUBY-1335 Extract Normalizer classes
The conditions under which we use either Encoding or Iconv based string
conversion don't change during the course of a single process. As such,
rather than checking the conditional repeatedly, use it to determine
which normalizer to use once, then hold onto that for the duration of
the application's run. Provided ~10% boost on the Marshalling perf
suite.