RUBY-1365 Limit the max depth of a backtrace that we'll aggregate
The JSON implementation in Ruby's stdlib cannot serialize JSON object graphs
that are nested deeper than a certain depth. The exact value varies by platform,
but this change places a sanity limit on the maximum nesting level in thread
profiles that the agent will produce.

Once you reach a nesting level this deep, the theory is that the leaf nodes
aren't really that interesting anyway.