RUBY-1241 Cap slow SQL queries
The list of Slow SQL queries was normalized by the agent, but no size
limit was applied to the in-memory storage. While we only send 10 to the
collector per harvest, we potentially hold onto vastly more than that.

This becomes a potential problem when either 1) the agent isn't
reporting for a very long time or 2) when an application has
sufficiently dynamic queries to bust our normalization logic in some way
(think constantly changing table names for instance).

Now for each query that comes in, we make sure that it's actually slow
enough to belong in the list that we will eventally send, and otherwise
don't hold onto it at all.