RUBY-1354 Cap max count of metrics in stasts hash
If we lose communication with the collector, or a client has a metric
explosion in progress, it's possible for us to gather vastly more
metrics than the collector will ever accept.

This change applies a limit to the number of unique metric specs that
will be taken by the stats hash. This prevents us from potential
problems with memory growth in those disconnected/exploding scenarios.