RUBY-1354 Cap metric count at merge
The StatsEngine#merge! is only called in two locations--when we are
merging back metrics from a failed call to the collector, or when we're
merging in data from a Resque child process.

We mainly care about the failed collector calls, and the check is
light-weight and unlikely to be hit under normal circumstances. Also for
the Resque case we're running off on another thread in the master
process, so it shouldn't impact transaction times for job execution.