RUBY-1290 Shorten to prefixes for Heroku dynos
Certain well known dyno names include a process ID and are thus too
dynamic for use a replacement hostname. In those cases, we want to clamp
down to just the prefix.

This is defaulted for now to "scheduler" and "run", but can be altered
by us if we find other names, or by users if they observe churn in their
hostnaming reported to us.

There is remaining tightening desired on the config side of how this is
handled, but it's waiting on another feature before that can be
completed.