RUBY-1290 Proper handling for string settings
Original code was only treating the Heroku dyno prefixes correctly if
they were sent as an array, which is only supported from YML.

This updates it to let basic CSV strings through, properly transforming
them before use. It also is more robust if we don't get either a String
or Array, defaulting to nothing for the setting and warning.

Once we have config transforms landed this should be revised to use
those, but since that won't make this release, we'll handle it manually
here.