RUBY-1354 Eliminate StatsMergerError
This error trap was added in an attempt to help track down RUBY-1159, but has
produced little in the way of usable signal since then. At the time it was
added, the merge! method on StatsHash was the primary path by which metrics from
a transaction were merged into the global StatsHash instance, but this is no
longer the case, meaning this trap isn't even really present on the interesting
code path anymore. We have a higher-level trap in Transaction#stop that should
capture these errors should they escape.