Bundling per Ruby version exclusive in multiverse
Planning to run multiverse suites concurrently on a single Ruby version
opens up problems that we've seen with bundling concurrently (in
particular when it's actually compiling a new native gem).

This commit applies some basic file locking so we only are bundling for
a given Ruby version one at a time (per user at least).