Don't rely on __FILE__ + FakeFS
When we changed from calling load to calling require on our multiverse
test files, this changed how __FILE__ was set. With load it was the
fully qualified path, with require it was now the relative path on
1.8.7, but not later Rubies.

When feeding these paths directly to FakeFS, things stopped lining up
for it's basic not-actually-a-full-file-system nature. Thus the tests
broke.

Thankfully, we were already capturing the actual multiverse directory
prior correctly with Dir.pwd, so just using that in place of __FILE__
for our location setting worked like a charm.