RUBY-1379 Labels initial/final whitespace bug
Our stripping only handled whitespace around our separator characters (:
and ;)  This meant that any leading/final spaces on a string would get
left there.

We now strip those to begin with from the string, then split and
individually strip each key and value as they come through, rather than
trying to do it in one big whack with a regex.

Note that this also pulls in the updates from the cross_agent_tests that
revealed this problem to begin with.