Author and Ruby version updates to newrelic_rpm
Wasn't aware of it, but there's a field in the gemspec for supported
Ruby version. As we look to eventually deprecate 1.8.7, it's a good idea
for us to properly note now what we support so we can change it in the
future.

Also updated the author list and re-applied alphabetic ordering based on
last name because, well, the alphabet.