RUBY-1345 Stay on protocol 12 until collector whitelist serialization is fixed
The collector is currently unable to properly serialize segment whitelist terms
to pruby. This manifests as an HTTP 200 response to /connect calls made to apps
that have these rules in place and are using pruby and protocol 14.

Even if this gets fixed in the collector, it won't be released before we want to
ship 3.9.5, so we're holding back on protocol 12 for now.

All of the machinery and tests for protocol 14 support will remain present, so
we can switch back whenever the collector is read.