RUBY-1380 Use JSON.generate instead of JSON.dump
The stdlib version of JSON.dump calls JSON.generate, which yajl will
override (when asked to). This means that our measures to try and use
the stdlib version end up still using yajl here, which we want to avoid
given the depth issues it's presented for our thread profiles.

Additionally this adds a multiverse test to confirm that we're not
touching the Yajl encoder, which is the core class it uses to generate
JSON.