RUBY-1380 Don'y rely on dump_default_options
The default options that we were passing to JSON.generate were only a
relatively recent addition to the json gem, so don't rely on them in
this way.

Thankfully the defaults were very stable before that, so we use the same
values that JSON.dump would have passed before dump_default_options was
added (json gem 1.6.2), and then use the defaults after that.

This also adds a full suite where we force the marshaller to JSON and
painstakingly walk through all the json gem versions from the earliest I
found we had reporting in (1.1.2) to present. Good fun.