Use Bundler.bundle_path to place lock file
The RUBY_DESCRIPTION for some Rubies (which we were using to construct the path
to a lock file) contains forward slashes, which have special meaning in
filesystem paths.

Instead of rolling our own lock file naming scheme, just drop a lock file in the
actual bundle_path that Bundler is going to install things to.