RUBY-1382 Fix for potential CAT race condition
If an HTTP call is being wrapped in a Timeout block, we can expect to
receive Timeout::Error at any given point in the code. Turns out,
there's actually a spot between a begin and the first line in the
contained block where we can raise, so setting our start time within a
begin/ensure doesn't actually guarantee that we'll get the time set.

Setting the time outside makes sure that if we reach the begin/ensure,
we do have a time to work with.