RUBY-1384 Perf improvement in RUM middleware
In the course of applying another fix, I noticed that we were calling to
NewRelic::Agent.browser_timing_header twice--once to see whether we
should be doing instrumentation, and then separately later on when
actually injecting it.

This is actually a pretty heavy method, with lots of allocations and
checks to it, so we should hold onto the value we got the first time to
use for actually injecting.