RUBY-1354 Never mind on metric limiting
Limiting the number of unique metrics stored in memory was a noble idea, but
turns out to be more trouble than it's worth, at lest for now. The previously
chosen limit of 2000 metrics / process / harvest was based on the PHP agent's
implementation, but the PHP agent also has a mechanism to 'force' recording of
important metrics that are essential to the proper functioning of the UI, which
we don't have. Without this mechanism, the cap of 2000 metrics turned out to be
too small for rpm_site, and may have been too small for our customers as well.

Given the performance impact of the metric limiting, the uncertainty about how
to set the limit, the fact that PHP is the only other agent that does this, and
the fact that we've never actually encountered a problem with unbounded metric
growth in the past, we're going to revert this change for now.