RUBY-1362 Transactions for ActiveJob
When we hook an ActiveJob's perform, we need to trigger a transaction.
This is either the first transaction (for job processors that we don't
instrument directly), or a nested transaction for those we do (Sidekiq,
Resque and DJ).

In the case of already instrumented processors, it's actually critical
that we DO start the nested transaction, because otherwise the
transactions get named after their ActiveJob wrapper classes, not the
user's job name itself.

Note that we do not record the MessageBroker consume metric in the case
where we do spin up a full transaction on perform, since it would be
duplicative of the main transaction time. If we're processing inline, we
will trace the consume, though, so the job processing shows up in the
transaction breakdown.