Don't set request to false if not found
If we're using ControllerInstrumentation in a setting where we don't
have a request object available (i.e. a background transaction), there's
code that is supposed to gracefully skip setting the transaction
options' request key. However, when there isn't a request this was
instead being set to false.

This caused downstream issues with some methods such as
TransactionState#is_request_transaction? which is expecting the
transaction request to be nil. While those uses aren't great either
and I'm going to clean them up, it's best if the request always be
either a full request object or nil, not a boolean.