RUBY-1362 Transaction categorization fix
TransactionState had two methods, in_background_transaction? and
in_request_transaction? which were used by thread profiling to
categorize what type of activity a given thread was using. This made the
poor choice of using the presence of transaction.request to determine if
we're web or not, which was unfortunately incorrect for background
transactions.

The more standard way of determining the type of transaction is already
on Transaction#record_web_transaction? which checks the category
assigned. We use that now instead for the TransactionState helpers
(although I've left them there since they neatly simplify the check if
you've already retrieved a TransactionState object).

TransactionState#is_web_transaction? is now also used from the ActiveJob
instrumentation when it's deciding whether the existing transaction is
web (in which case we don't start a nested transaction since we're
inline with a web request and don't want to alter the transaction name).