Tighten up checking on build stage
Noticed while testing out some changes to scripts that later versions of
Rake end up with an empty string instead of a nil for the build stage.
This causes the gem build to fail with an invalid version format.

Not actually updating Rake at this point, but being more conservative
we'll be set if/when we choose to move that dependency forward.