Use config transforms on dyno prefixes
The list of Heroku dyno prefixes to shorten was previously hand-rolling
the code to process a config value with a CSV list in it. With the new
config transforms that isn't necessary anymore.

This also makes a slight change to DefaultSource#convert_to_list so it
will raise explicitly on finding a value it couldn't make a list from.
This is caught by the usage upstream in the config manager, and gives
the expected errors in the log for badly formatted config values.