Extend ActiveRecord multiverse to other versions
We were only running the (currently very slim) ActiveRecord multiverse
suite against Rails 3.2. Not much point in that level of coverage.

In the future it'd be nice to see the unit tests against ActiveRecord
move over into this suite now that it covers all the versions.

Additionally this sets things up for creating a multiverse suite against
DelayedJob using the ActiveRecord adapter.