RUBY-1389 Provide named queue metrics where possible
There was dead code that appears to have never been running for writing
metrics about the DJ queue name as well as priority. Since we're going
to be testing the sampler, now seemed like a good time to fix this.

We will now include metrics titled
Workers/DelayedJob/queue_length/name/<queue_name> where the last bit is
the name of the job's queue, or "default" if there isn't one given.

Older DJ versions didn't include the queue, so on those the metric is
not recorded. Priority metrics are recorded across all versions we
support for DJ.