RUBY-1391 Capture synthetics TT's
In addition to all our other ways of capturing TT's, we will capture up
to 20 TT's that are synthetics requests, regardless of their duration.

This also applies the custom parameters that are requested to the TT
that we do capture (although it doesn't yet change the top-level
final parameter we're supposed to add to the array).