Only watch our threads
Our thread watcher was occasionally seeing non-Ruby (or at least not
our) threads when running on JRuby. These show up without a Hometown
trace on their instantiation, which suggests they are JVM/JRuby system
threads of some variety that we don't care about.

Bottom line, we're only trying to count the type of threads that
Hometown can see created, so limit ourselves to that.