RUBY-1394 Remove BrowserToken, and ttGuid from RUM config
The BrowserToken class was previously part of a complex system to allow the
beacon to trigger and preserve app-server transaction traces, via the NRAGENT
HTTP cookie.

With the advent of Browser Session Traces, this system is no longer necessary,
so its baggage can be removed. The inclusion of the ttGuid in the RUM config was
also part of this system, and has been removed as well.

This change allows us to remove a piece of state from TransactionState, and stop
parsing HTTP cookies on every request (woohoo!).