RUBY-1387 Just use one buffer for all custom events
We previously had a scheme in place (mainly to support Grabby's use case)
whereby you could register separate event buffers with different capacities by
name. Given that we're no longer planning to expose this via the API, and Grabby
is still a ways off, the added complexity of this scheme isn't worth it, so this
change brings us back to having only one buffer for all custom event types.