RUBY-1391 Remove InboundRequestMonitor#from_headers
This method was intended to be used to extract header values from incoming Rack
env hashes. It dealt with keys in the Rack env hash that were cased differently
from the keys being searched for, and also was used to search for multiple
possible key names (with and without the 'HTTP_' prefix).

In practice, it appears that none of this is necessary. By spec, HTTP header
names will always be uppercased, have '-' converted to '_', and be prefixed with
'HTTP_' during the construction of the Rack env hash.

Many tests that used Rack::Test or ActionDispatch::IntegrationTest and specified
custom request headers had to be updated to work with this change: both of these
APIs expect that any custom headers passed into the 'get' or 'post' (etc)
methods will be merged *directly* into the Rack env hash, rather than going
through the usual conversion process as would happen with a real web server.