RUBY-1395 Do not shutdown the agent after each Resque job
Instead, just flush any collected data on the pipe back to the parent process.
This is important because with 3rd-party Resque forking hacks like
resque-multi-job-forks and resque-jobs-per-fork, or when Resque is run with the
FORK_PER_JOB environment variable set to false, we will end up processing
multiple jobs in the same process, and we don't want to shutdown the agent after
the first one.