RUBY-1406 cache SSL cert store
Instead of re-parsing all 47 certs in our cacert.pem file every time we need to
establish a new connection to the collector, this change makes it so that we
cache the parsed CA bundle in an OpenSSL::SSL::Store object hanging off of the
NewRelicService object.

The store will be lazily created, and only re-generated if the requested cert
file path changes (e.g. because of a config change).