RUBY-1408 Config lookup with warbler
When using warbler on JRuby, a Ruby app gets packaged up into a single
jar or war file. When that happens, we're not looking in the right
location for the config file.

GEM_HOME in the environment is set to the base path where we should go
hunting for our configs, so if we find that we're in a jar'd location,
tweak our search paths.