RUBY-1285 Factor in seen? check more consistently
Although I'm not sure we want tag_as_seen as a truly public interface,
it was awkward that it was only being applied partially, and as part of
the metric counting, in the error collector.

This revises that check to happen at the higher level where we choose to
skip, and cleans up some of how we do the thread local access and
tagging.

Note that under normal circumstances, the queue itself prevented us from
getting duplicate traces recorded, so it's only in the odd case of
someone calling tag_as_seen directly that it would be revealed that we
weren't abiding by it.