RUBY-1409 Another shot for rejected synthentics
If we reject a synthetics event from the buffer, it is given a chance to
live on in the main transaction events buffer.

Rejection for a synthetics event is now based both on size of the buffer
and the timestamp of the event, so we retain the first events (even on a
remerge).