Don't mutate config from within in_transaction helper
The in_transaction helper is used to simulate a transaction in many of our
tests, independent of specific framework instrumentation. Because a transaction
may kick-start the harvest thread, and we don't want the harvest thread to spawn
during our tests, we had previously added a config layer from within
in_transaction to set disable_harvest_thread to true.

This turns out to be problematic for tests that simulate transactions from
multiple threads, because the config hash has no synchronization around it. This
led to failures where multiple threads were attempting to mutate the config hash
at the same time.