RUBY-1420 Don't use :grape dependency name
The newrelic-grape gem uses a dependency detection block with the same
name (:grape) that we used in our currently disabled instrumentation.
The agent doesn't allow for double-registration of instrumentation
blocks, so this causes clients with 3.9.8 + newrelic-grape to stop
getting instrumentation.

For the next release, we'll just rename our key. I have a plan for how
we'll handle this more gracefully for the real Grape instrumentation
release.