RUBY-1418 Improve Sinatra at_exit behavior
This tightens the conditions under which we'll skip installation of the agent's
at_exit hook. Previously, we would skip installation of this hook if we found
that the Sinatra::Application constant was defined.

This was done because in some ways of running Sinatra classic apps, they'll run
entirely from an at_exit handler. If we install our at_exit handler after
Sinatra has already installed its, ours will run first, and we'll incorrectly
shut down the agent before it has had the chance to collect any data.

Just checking for the presence of the Sinatra::Application constant turns out to
have been far too loose a condition - it triggers in many contexts where Sinatra
happens to be used incidentally, and prevents the at_exit hook from firing when
we need it to in these contexts.

Instead, we'll now mimic the check that Sinatra itself does in order to
determine whether to run the app from an at_exit hook.