RUBY-1404 Moved frame naming from set default / set overridding name methods
This is only a slight a improvement, but the methods set_default_transaction_name
and set_overriding_transaction_name previously set three different values.  We
ultimately would like it to set only one value as the method name indicates, but
further work needs to be done to decouple category from name.