RUBY-1404 Don't try to convert Hashie::Mash to a Hash before filtering
Doing this turns out to be harmful in the case where we're on Ruby 1.9.3,
because Hash#to_h doesn't exist on that Ruby version, and Hashie::Mash doesn't
define its own to_h method.

The parameter filtering methods do not mutate the original params hash that is
passed in, so passing in the raw Hashie::Mash instance should be fine.