Fix for Time.now stubbing failures
We periodically see failures in the Minitest runner calling Time.now
after tests where we've frozen time by stubbing out Time.now. This is
supposed to get cleaned up by the Mocha integration teardown, but under
some circumstances doesn't seem to be.

Since this is a pretty isolated and rare case (something we'd stub that
Minitest itself would also use), just provide an explicit means to reset
it by executing with a block.