Only pull in Guard on 1.9.x+
Ran into a problem as yet another dependency for Guard moved to a spot
where it was incompatible with Ruby 1.8.7.

Given that we don't do primary development on Ruby 1.8.7, this commit
moves Guard to be conditional on using the agent with a new-enough Ruby.
This lets us get back onto the latest and greatest Guard.