RUBY-1404 Name Grape segments more meaningfully
In our Grape instrumentation, we would previously name the segment representing
the call method on the Grape::API subclass after the transaction as a whole.

This created strange names in the UI when looking at the breakdown chart + table
for a Grape app. You would see a breakdown component call something like
'MyGrapeApiClass /users/:id (GET)'. Seeing this in the breakdown is super
confusing, because its name is identical to the name of the transaction as
a whole.

Instead, with this change, we will name the segment representing the call method
on the Grape::API subclass as 'Middleware/Grape/MyGrapeApiClass/call', which
will be presented in the UI as 'MyGrapeApiClass#call' (with a category of
'Grape').

The name of the transaction as a whole will not change.