Re-implement freeze/unfreeze_time outside Mocha to make them thread-safe
Something in Mocha's unstub method appears to be non-atomic, meaning that if
a background thread spawned during a test happens to call Time.now at the wrong
moment, while we're in the midst of tearing down the stub, we get test failures.
This change re-implements our freeze_time / advance_time / unfreeze_time helpers
without relying on Mocha in such a way that changes to the Time class should be
atomic.