RUBY-1439 Datastore metrics for ActiveRecord
Change over recording of metrics from ActiveRecord/* style to the new
Datastore metric namespacing.

Currently only sets the Product as ActiveRecord rather than extracting
from the config. Will fix that in a separate commit.

In addition to that change, this also removes the RemoteService metrics
which were only used internally and won't be recorded by the agent any
more at this time.