RUBY-1440 Truncate needs to mutate database statement since it subclasses string
The work here will probably end up being undone in a later commmit.  The
NewRelic::Database::Statement subclasses string, which is generally a bad idea
and has caused pain in attempting to be able to append changes to it when there
are multiple calls to notice_sql during a stack frame as it can only be mutated,
at any point that it is assigned to a new string the database statement is clobbered.
The correct fix is to not subclass string in NewRelic::Database::Statement.