Another fix for thread profiling test timeouts
On JRuby, we occasionally see long pauses in CI which cause the thread
profiling tests to hit our timeout condition while waiting for
sufficient profiling to occur.

As it turns out, we were trying to wait for 10 iterations on the
profiling loop, but in fact can get away with a lot fewer. Based on
diagnostic output, we are getting at least one cycle through, even in
the stalled tests, and the logic only requires one, so setting to just
hop out after a single iteration might do the trick.