RUBY-1437 Fix for passing wrong scoped metric
Similar problem to what we had in ActiveRecord instrumentation,
trace_execution_scoped was modifying the array that we were subsequently
using. Because these API's are new, and we only really needed the scoped
metric, I just changed them to take that single metric instead of the
full list which avoids unnecessary allocation.