Confirm unsupported on Mongo 2.x
The new version the mongo gem has been released, which is a full rewrite
and not compatible with 1. which we instrumented. We added explicit
checks at the time for Mongo 2.x to make sure we don't try to instrument
it, but without the gem we couldn't easily write a multiverse test.

Now that the gem is available, apply the standard unsupported check that
we were doing against oldest 1.x versions to 2.x as well. Should be able
to query without a problem and without us generating metrics.