RUBY-1415 Stop using hashes with NoticedError
The primary interface to NoticedError via the ErrorCollector has been a
series of mangled hashes that get merged up from user input and other
locations. This did WAY more allocations than we want and was super hard
to follow.

This commit changes that to use direct access for the known fields on
NoticedErrors that we support, which simplifies things greatly and cuts
down on the memory usage when noticing an error.