RUBY-1415 Rename Transaction#uri => Transaction#request_path and simplify Transaction#user_defined_rules_apply?
Changes in 6d4eced695bdfc8872d5a0fd25db77b7dc966066 ensured that we what were storing in
Transaction#uri was a path.  This commit changes Transaction#uri to refer to request_path.
This also allowed us to simplify the implementation of Transaction#user_defined_rules_apply?
as it is no longer necessary to do heavy processing of a uri to strip off username, password,
and query string.  We were also able to delete a number of tests that were testing that transactions
are ignored for a number of (now) impossible formats for request_path.