RUBY-1415 Clean up ServerSource, explicit whitelisting for SSC keys
There is a new required key for each entry in the DEFAULTS hash called
:allowed_from_server, which should be true or false, and specifies whether the
key will be respected if received as part of server-side config.

In addition, we now whitelist the keys from the connect reply that are actually
allowed to be merged into the config hash, instead of blindly merging everything
from both the top level and the nested agent_config hash in.