Remove ShimAgent in favor of inline checks in public API methods
Also removes several other *::Shim modules that were used from the ShimAgent.
The shim pattern was intended to deal with cases where the agent code was
loaded, and APIs on the agent might have been called, but agent_enabled was
false, meaning that the agent's harvest thread wouldn't be running, and
instrumentation wouldn't be installed.

It took a rather haphazard and poorly-tested approach to shimming out bits of
the agent's guts. It hasn't been maintained for the past several years, and no
longer feels like the right abstraction to solve this problem.

Instead, we're moving to a model where we clearly define what our public API is,
and ensure that public API methods are safe to call even when the agent is
disabled. Calls to non-public APIs when the agent is disabled might fail in
strange ways, but the only entrypoints for those calls should be either a) via
public APIs, which will short-circuit, or b) via instrumentation, which won't be
installed when the agent is disabled.