RUBY-1470 Tag seen errors to avoid double-counting them
When the error collector sees an error, we'll now tag it with a special instance
variable in order to avoid counting the same error again.

This works in concert with our de-duping of exceptions that happen within
a transaction, which works by storing exceptions during the transaction in
a hash keyed by the exception objects.

We've used this approach before, and ran into trouble with frozen objects (see
RUBY-1114) and JRuby wrapped Java objects (see RUBY-1153). This time around,
we're putting in exceptions for both of those cases and hoping that the
per-transaction de-duping will cover us when those come up.