RUBY-1480 Ensure at_exit installed
JRuby by default doesn't install at_exit (for historical reasons that
aren't likely to be true). Because it's critical for rake
instrumentation, make sure that we give a call to shutdown either way.

If at_exit's do get installed generally for the agent, this might be
unnecessary, but is safe and fast to call when disconnected.