RUBY-1480 Add "tasks/newrelic" for Rails + Rake
If a Rails Rakefile wants to instrument tasks that don't rely on
:environment, the agent won't be loaded as necessary when the tasks are
run. This change adds 'tasks/newrelic' which may be required in your
Rakefile to kickstart the agent in these environments.

This should not be required outside of Rails, or if you're instrumenting
tasks that do rely on loading the full Rails environment.