RUBY-1485 Fix for model naming in nested datastore calls
If we apply an ActiveRecord model name via with_database_metric_name and
a model ends up generating a call to another instrumented datastore
(i.e. validation hitting Memcached), the inner datastore call would get
annotated with the model name.

This is not desirable in most cases, so we now provide a mechanism for
constraining your database name overrides to your product.