RUBY-1498 Port mongo operation tests to mongo2_instrumentation_test
Ported the mongo operation tests to the mongo2 instrumentation test
which tests the metrics recorded by the mongo command subscriber.
Presently there are a number of cases where metrics recorded by the
subscriber differ from what is expected based on previous versions
of instrumentation.  In some of the cases the expectations have
been modified to make the tests pass, but comments were inserted
noting the discrepancies.  In other cases the tests are failing
due to other inconsistencies that require larger changes.  Those
tests are currently being skipped and are annotated with summaries
of what needs to be changed to get them passing.