RUBY-1498 Update tests to fail
In the initial pass for the mongo multiverse I adjusted assertions
or used the skip method to make tests pass. In this commit I've
changed the assertions to reflect what we ultimately want from
the instrumentation. Many of these tests are failing, but making
them pass will drive the instrumentation in the proper direction.