RUBY-1501 Allow SampledBuffer#append to accept a block
This is slightly unfortunate as it leads to an additional allocation
per sample added through append with a block, but is necessary to
conform to the EventBuffer interface where append increments @seen and
calls append_event which is a template method that subclasses are expected
to define. Previously we were able to avoid that allocation by adding items
by calling append_event directly with an implicit block, however that was
incorrect as it failed to increment the number of samples seen.