RUBY-1501 Send error events on harvest
This adds the error event aggregator as a container to be harvested
during the main harvest cycle. There was some discussion about
putting it on it's own period, and this can be done if we decide we
want it. Some modifications to harvest were necessary to support
containers that contain metadata. For now it is only the ErrorEventAggregator, but
it will soon include the TransactionEventAggregator and CustomEventAggregator.