RUBY-1501 Move error trace related data container methods from error collector
During the intial extraction of the ErrorTraceAggregator from the ErrorCollector
the data container methods harvest!, merge!, and reset! were left intact and
modified to delegate to the ErrorTraceAggregator.  This PR moves those methods
from the ErrorCollector to the ErrorTraceAggregator and updates references
in the rest of the codebase.  Additionally, tests related to the ErrorTraceAggregator
were moved from the ErrorCollectorTest to the ErrorTraceAggregatorTest.