RUBY-1501 Handle enabling or disabling of Error Traces / Events
The logic regarding whether or not to process a noticed error has
become slightly more difficult as it needs to check if either
the ErrorTraceAggregator or ErrorEventAggregator is enabled. The
ErrorCollector will now short circuit in notice_error if both
buffers are disabled, and each buffer will also short circuit
on methods that add new items when they are disabled.