RUBY-1501 Add new request agent attributes
This is a first take at adding new request attributes from the updated
agent attributes spec. Code touching the transaction is a hot path and
as such we worry about allocations.  This implementation is pretty ugly,
but is going to be the one that incurs the least number of allocations.
The Rack::Request#host method is a source of the majority of allocations
incurred by these changes as a single call results in 8 allocations. I
will followup this commit with an implementation that is slightly less
ugly and evaluate if it is worth the allocations or not.

Here are the performance changes from this commmit:

wear:ruby_agent mwear$ ./test/performance/script/runner -n test_basic_middleware_stack -C
1 tests, 0 failures, 7.550711 s total
+--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+
| name                                       | before    | after     | delta | allocs_before | allocs_after | allocs_delta |
|--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------|
| RackMiddleware#test_basic_middleware_stack | 429.58 µs | 414.91 µs | -3.4% |           244 |          252 |        +3.3% |
+--------------------------------------------+-----------+-----------+-------+---------------+--------------+--------------+